+++ packages/docs/docs/main/change-log.md
@@ -30,6 +30,9 @@ This page lists the changes to the IsaacScript framework.
   - `UppercaseKeys`
   - `StartsWithLowercase`
   - `StartsWithUppercase`
+- Added the following decorator functions:
+  - `@Callback` (for specifying that a class method should automatically be subscribed to the corresponding callback)
+  - `@CallbackCustom` (for specifying that a class method should automatically be subscribed to the corresponding custom callback)
 
 ## September 22th, 2022